Fixing _REV_INVALID_HART_ID_ upcasting to 0xFFFFFFFF
There was an issue where HartToExec would be equal to an invalid ID (0xFFFF) however when compared to _REV_INVALID_HART_ID_ (0xFFFFFFFF) wouldn't be flagged as invalid.

+++ include/RevInstTable.h
@@ -34,7 +34,7 @@
 #endif
 
 #ifndef _REV_INVALID_HART_ID_
-#define _REV_INVALID_HART_ID_ ~(uint16_t(0))
+#define _REV_INVALID_HART_ID_ (uint16_t)~(uint16_t(0))
 #endif
 
 // Masks
